Move ConnectionTable and related functions into its own module
