+++ ouroboros-network/src/Ouroboros/Network/NodeToClient.hs
@@ -121,7 +121,7 @@ connectTo =
 -- | A specialised version of @'Ouroboros.Network.Socket.withServerNode'@
 --
 withServer
-  :: ConnectionTable
+  :: ConnectionTable IO
   -> Socket.AddrInfo
   -> (Socket.SockAddr -> Socket.SockAddr -> peerid)
   -- ^ create peerid from local address and remote address
